work

+++ src/SkewDetector.h
@@ -47,7 +47,7 @@ public:
 	/** probality measure 1 */
 	int probMeasure1;
 	/** probality measure 2 */
-	double probMeasure2;
+	int probMeasure2;
 };
 
 /**
